:bug: Fix gradient and scroll flash on auth and onboarding navigation
